Disable thread actions for users that have not joined the community (#4521)
* Quill initial work

* Add monospace font on markdown

* Add focus styles

* Add disabled state

* prevent jump between markdown toggle

* Adjust quill all over the app

* Update icons to phosphor

* Add quill to component showcase

* Remove placeholder default text

* Add banner component

* Add static banner to view thread page

* Add static banner to create thread page

* Add closing banner in new thread form

* Add closing banner in discussion page

* Abstract join community logic into hook

* use join community hook in new thread form

* use join community hook in view thread page

* use hook for check active account

* use hook for closing banner logic

* abstract AccountSelectorModal

* abstract TOS modal

* abstract all join community modals modals

* rename old banner and create new component for join community banner

* Fix ts errors

* Revert merge changes

* Disable editor 4377

* Change naming

* Change naming pt. 2

* Change naming pt. 3

* Unify small upvote button

* Remove upvote from thread action component

* Replace comment button

* Replace share button

* Replace subscribe btn

* Replace overflow btn

* add upvote to thread action component

* Change div to button

* Progress on disabling

* Add tooltip wrapper to thread action

* Edit tooltip styles to adjust new designs

* Add tooltip to disabled buttons

* Handle reply button

* Fix create thread buttons

* self CR

* fixed lint errors

* remove comment

* conflicts

---------
