+++ packages/commonwealth/client/scripts/views/components/ReactionButton/CommentReactionButton.tsx
@@ -1,6 +1,5 @@
-import 'components/ReactionButton/CommentReactionButton.scss';
-import { notifyError } from 'controllers/app/notifications';
 import React, { useState, useEffect } from 'react';
+import { notifyError } from 'controllers/app/notifications';
 import app from 'state';
 import type ChainInfo from '../../../models/ChainInfo';
 import type Comment from '../../../models/Comment';
@@ -8,45 +7,41 @@ import ReactionCount from '../../../models/ReactionCount';
 import {
   useCreateCommentReactionMutation,
   useDeleteCommentReactionMutation,
-  useFetchCommentReactionsQuery
+  useFetchCommentReactionsQuery,
 } from '../../../state/api/comments';
 import Permissions from '../../../utils/Permissions';
 import { LoginModal } from '../../modals/login_modal';
-import { CWIcon } from '../component_kit/cw_icons/cw_icon';
 import { Modal } from '../component_kit/cw_modal';
-import { CWTooltip } from '../component_kit/cw_popover/cw_tooltip';
-import {
-  getClasses,
-  isWindowMediumSmallInclusive,
-} from '../component_kit/helpers';
-import {
-  getDisplayedReactorsForPopup,
-  onReactionClick,
-} from './helpers';
+import { isWindowMediumSmallInclusive } from '../component_kit/helpers';
+import { getDisplayedReactorsForPopup, onReactionClick } from './helpers';
+import CWUpvoteSmall from 'views/components/component_kit/new_designs/CWUpvoteSmall';
 
 type CommentReactionButtonProps = {
   comment: Comment<any>;
+  disabled: boolean;
 };
 
 export const CommentReactionButton = ({
   comment,
+  disabled,
 }: CommentReactionButtonProps) => {
-  const [isLoading, setIsLoading] = useState<boolean>(false);
   const [reactors, setReactors] = useState<Array<any>>([]);
   const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
   const [reactionCounts, setReactionCounts] = useState<ReactionCount<any>>();
-  const { mutateAsync: createCommentReaction } = useCreateCommentReactionMutation({
-    commentId: comment.id,
-    chainId: app.activeChainId()
-  })
-  const { mutateAsync: deleteCommentReaction } = useDeleteCommentReactionMutation({
-    commentId: comment.id,
-    chainId: app.activeChainId()
-  })
+  const { mutateAsync: createCommentReaction } =
+    useCreateCommentReactionMutation({
+      commentId: comment.id,
+      chainId: app.activeChainId(),
+    });
+  const { mutateAsync: deleteCommentReaction } =
+    useDeleteCommentReactionMutation({
+      commentId: comment.id,
+      chainId: app.activeChainId(),
+    });
   const { data: reactions } = useFetchCommentReactionsQuery({
     chainId: app.activeChainId(),
     commentId: comment.id,
-  })
+  });
 
   useEffect(() => {
     const redrawFunction = (comment_id) => {
@@ -82,8 +77,6 @@ export const CommentReactionButton = ({
       return r.Address.address === activeAddress;
     });
 
-    setIsLoading(true);
-
     deleteCommentReaction({
       canvasHash: foundReaction.canvas_hash,
       reactionId: foundReaction.id,
@@ -92,34 +85,49 @@ export const CommentReactionButton = ({
         likes: likes - 1,
         hasReacted: false,
       },
-    }).then(() => {
-      setReactors(
-        reactors.filter(({ Address }) => Address.address !== userAddress)
-      );
-      setReactionCounts(app.comments.reactionCountsStore.getByPost(comment));
-      setIsLoading(false);
-    }).catch(() => {
-      notifyError('Failed to update reaction count');
-    });
+    })
+      .then(() => {
+        setReactors(
+          reactors.filter(({ Address }) => Address.address !== userAddress)
+        );
+        setReactionCounts(app.comments.reactionCountsStore.getByPost(comment));
+      })
+      .catch(() => {
+        notifyError('Failed to update reaction count');
+      });
   };
 
   const like = (chain: ChainInfo, chainId: string, userAddress: string) => {
-    setIsLoading(true);
-
     createCommentReaction({
       address: userAddress,
       commentId: comment.id,
       chainId: chainId,
-    }).then(() => {
-      setReactors([
-        ...reactors,
-        { Address: { address: userAddress, chain } },
-      ]);
-      setReactionCounts(app.comments.reactionCountsStore.getByPost(comment));
-      setIsLoading(false);
-    }).catch(() => {
-      notifyError('Failed to save reaction');
     })
+      .then(() => {
+        setReactors([
+          ...reactors,
+          { Address: { address: userAddress, chain } },
+        ]);
+        setReactionCounts(app.comments.reactionCountsStore.getByPost(comment));
+      })
+      .catch(() => {
+        notifyError('Failed to save reaction');
+      });
+  };
+
+  const handleVoteClick = async (e) => {
+    e.stopPropagation();
+    e.preventDefault();
+
+    if (!app.isLoggedIn() || !app.user.activeAccount) {
+      setIsModalOpen(true);
+    } else {
+      onReactionClick(e, hasReacted, dislike, like);
+    }
+  };
+
+  const handleVoteMouseEnter = async () => {
+    setReactors(reactions);
   };
 
   return (
@@ -130,66 +138,16 @@ export const CommentReactionButton = ({
         onClose={() => setIsModalOpen(false)}
         open={isModalOpen}
       />
-      <button
-        className={getClasses<{ disabled?: boolean }>(
-          { disabled: isLoading || isUserForbidden },
-          `CommentReactionButton ${hasReacted ? ' has-reacted' : ''}`
-        )}
-        // TODO: won't be needed now?
-        onMouseEnter={async () => {
-          setReactors(reactions);
-        }}
-        onClick={async (e) => {
-          e.stopPropagation();
-          e.preventDefault();
-
-          if (!app.isLoggedIn() || !app.user.activeAccount) {
-            setIsModalOpen(true);
-          } else {
-            onReactionClick(e, hasReacted, dislike, like);
-          }
-        }}
-      >
-        {likes > 0 ? (
-          <CWTooltip
-            content={
-              <div className="reaction-button-tooltip-contents">
-                {getDisplayedReactorsForPopup({
-                  reactors: reactors.map((r) => r.Address.address),
-                })}
-              </div>
-            }
-            renderTrigger={(handleInteraction) => (
-              <div
-                onMouseEnter={handleInteraction}
-                onMouseLeave={handleInteraction}
-                className="btn-container"
-              >
-                <CWIcon
-                  iconName="upvote"
-                  iconSize="small"
-                  {...(hasReacted && { weight: 'fill' })}
-                />
-                <div
-                  className={`reactions-count ${hasReacted ? ' has-reacted' : ''
-                    }`}
-                >
-                  {likes}
-                </div>
-              </div>
-            )}
-          />
-        ) : (
-          <>
-            <CWIcon iconName="upvote" iconSize="small" />
-            <div
-              className={`reactions-count ${hasReacted ? ' has-reacted' : ''}`}
-            >
-              {likes}
-            </div>
-          </>
-        )}
-      </button>
+      <CWUpvoteSmall
+        voteCount={likes}
+        disabled={isUserForbidden || disabled}
+        selected={hasReacted}
+        onMouseEnter={handleVoteMouseEnter}
+        onClick={handleVoteClick}
+        tooltipContent={getDisplayedReactorsForPopup({
+          reactors: reactors.map((r) => r.Address.address),
+        })}
+      />
     </>
   );
 };
