+++ packages/commonwealth/client/scripts/views/pages/discussions/CommentCard/CommentCard.scss
@@ -1,4 +1,5 @@
 @import '../../../../../styles/shared';
+@import '../../../../../styles/components/component_kit/cw_popover/cw_tooltip';
 
 .thread-connectors-container {
   display: flex;
@@ -23,7 +24,6 @@
 .Comment {
   display: flex;
 
-
   &.highlighted {
     background-color: $yellow-200;
     border-radius: $border-radius-corners;
@@ -43,7 +43,6 @@
     gap: 12px;
 
     .comment-text {
-
       div,
       span,
       p {
@@ -94,63 +93,10 @@
     .comment-footer {
       display: flex;
       align-items: center;
-      gap: 8px;
+      gap: 1px;
 
-      .comment-option-btn {
-        outline: 0;
-        border: 0;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        gap: 4px;
-        cursor: pointer;
-        padding: 4px 8px;
-        border-radius: $border-radius-corners-wider;
-        color: $neutral-600;
-
-        &:hover {
-          background-color: $neutral-200;
-        }
-
-        .Text {
-          color: $neutral-600;
-        }
-      }
-
-      .reply-button {
-        align-items: center;
-        cursor: pointer;
-        display: flex;
-        gap: 4px;
-
-        .Icon,
-        .menu-buttons-text.Text {
-          color: $neutral-500;
-        }
-
-        &:hover {
-
-          .Icon,
-          .menu-buttons-text.Text {
-            color: $neutral-600;
-          }
-        }
-
-        &:active {
-
-          .Icon,
-          .menu-buttons-text.Text {
-            color: $neutral-800;
-          }
-        }
-
-        &.selected {
-
-          .Icon,
-          .menu-buttons-text.Text {
-            color: $neutral-800;
-          }
-        }
+      .Tooltip {
+        @include newTempBlackTooltip;
       }
     }
   }
@@ -158,14 +104,14 @@
 
 .CommentActions {
   .danger {
-    color: $rorange-600  !important;
+    color: $rorange-600 !important;
 
     .Icon {
-      color: $rorange-600  !important;
+      color: $rorange-600 !important;
     }
 
     .Text {
-      color: $rorange-600  !important;
+      color: $rorange-600 !important;
     }
   }
 }
