--- packages/commonwealth/client/scripts/views/pages/discussions/ThreadCard/Options/AdminActions/index.scss
@@ -1,36 +1,0 @@
-@import '../../../../../../../styles/shared';
-
-.AdminActions {
-  .danger {
-    color: $rorange-600  !important;
-
-    .Icon {
-      color: $rorange-600  !important;
-    }
-
-    .Text {
-      color: $rorange-600  !important;
-    }
-  }
-}
-
-.action-btn {
-  outline: 0;
-  border: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 4px;
-  cursor: pointer;
-  padding: 4px 8px;
-  border-radius: $border-radius-corners-wider;
-  color: $neutral-600;
-
-  &:hover {
-    background-color: $neutral-200;
-  }
-
-  .Text {
-    color: $neutral-600;
-  }
-}
