--- packages/commonwealth/client/scripts/views/pages/discussions/ThreadCard/Options/index.scss
@@ -1,28 +1,0 @@
-@import '../../../../../../styles/shared';
-
-.Options {
-  align-items: center;
-  display: flex;
-  gap: 8px;
-
-  .thread-option-btn {
-    outline: 0;
-    border: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 4px;
-    cursor: pointer;
-    padding: 4px 8px;
-    border-radius: $border-radius-corners-wider;
-    color: $neutral-600;
-
-    &:hover {
-      background-color: $neutral-200;
-    }
-
-    .Text {
-      color: $neutral-600;
-    }
-  }
-}
