+++ packages/commonwealth/client/scripts/views/pages/discussions/ThreadCard/ThreadOptions/AdminActions/AdminActions.scss
@@ -1,0 +1,15 @@
+@import '../../../../../../../styles/shared';
+
+.AdminActions {
+  .danger {
+    color: $rorange-600 !important;
+
+    .Icon {
+      color: $rorange-600 !important;
+    }
+
+    .Text {
+      color: $rorange-600 !important;
+    }
+  }
+}
