+++ packages/commonwealth/client/scripts/views/pages/discussions/ThreadCard/ThreadOptions/AdminActions/index.tsx
@@ -1,0 +1,4 @@
+import { AdminActions } from './AdminActions';
+import type { AdminActionsProps } from './AdminActions';
+
+export { AdminActions, AdminActionsProps };
