+++ packages/commonwealth/client/scripts/views/pages/discussions/ThreadCard/ThreadOptions/ReactionButton/ReactionButton.scss
@@ -49,51 +49,6 @@
       &.has-reacted {
         color: $primary-500;
       }
-
     }
   }
-}
-
-.CommentReactionButton {
-  outline: 0;
-  border: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 4px;
-  cursor: pointer;
-  padding: 4px 8px;
-  border-radius: $border-radius-corners-wider;
-  color: $neutral-600;
-  font-weight: 400;
-
-  &:hover {
-    background-color: $neutral-200;
-  }
-
-  &.has-reacted {
-    color: $primary-500;
-  }
-
-  .Text {
-    color: $neutral-600;
-  }
-
-  &.disabled {
-    pointer-events: none;
-  }
-
-  .menu-buttons-text.Text {
-    color: $neutral-600;
-    font-weight: 400;
-
-    &.has-reacted {
-      color: $primary-500;
-    }
-  }
-}
-
-.reaction-button-tooltip-contents {
-  display: flex;
-  flex-direction: column;
 }
