+++ packages/commonwealth/client/scripts/views/pages/discussions/ThreadCard/ThreadOptions/ReactionButton/index.tsx
@@ -1,0 +1,3 @@
+import { ReactionButton } from './ReactionButton';
+
+export { ReactionButton };
