+++ packages/commonwealth/client/scripts/views/pages/discussions/ThreadCard/ThreadOptions/ThreadOptions.scss
@@ -1,0 +1,11 @@
+@import '../../../../../../styles/components/component_kit/cw_popover/cw_tooltip';
+
+.ThreadOptions {
+  align-items: center;
+  display: flex;
+  gap: 1px;
+
+  .Tooltip {
+    @include newTempBlackTooltip;
+  }
+}
