+++ packages/commonwealth/client/scripts/views/pages/discussions/ThreadCard/ThreadOptions/index.tsx
@@ -1,0 +1,3 @@
+import { ThreadOptions } from './ThreadOptions';
+
+export { ThreadOptions };
