--- packages/commonwealth/client/styles/components/ReactionButton/CommentReactionButton.scss
@@ -1,55 +1,0 @@
-@import '../../shared';
-
-.CommentReactionButton {
-  outline: 0;
-  border: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 4px;
-  cursor: pointer;
-  padding: 4px 8px;
-  border-radius: $border-radius-corners-wider;
-  color: $neutral-600;
-  font-weight: 400;
-
-  &:hover {
-    background-color: $neutral-200;
-  }
-
-  &.has-reacted {
-    color: $primary-500;
-  }
-
-  .Text {
-    color: $neutral-600;
-  }
-
-  &.disabled {
-    pointer-events: none;
-  }
-
-  .menu-buttons-text.Text {
-    color: $neutral-600;
-    font-weight: 400;
-
-    &.has-reacted {
-      color: $primary-500;
-    }
-  }
-
-  .reactions-count {
-    font-size: 12px;
-  }
-
-  .btn-container {
-    display: flex;
-    gap: 4px;
-    align-items: center;
-  }
-}
-
-.reaction-button-tooltip-contents {
-  display: flex;
-  flex-direction: column;
-}
