+++ packages/commonwealth/client/styles/components/component_kit/new_designs/cw_thread_action.scss
@@ -8,20 +8,49 @@
   gap: 4px;
   width: fit-content;
   color: $neutral-500;
-}
+  border: 0;
 
-.default {
-  color: $neutral-500;
-}
+  .Text {
+    color: $neutral-500;
+  }
 
-.hover {
-  background-color: $neutral-100;
-  color: $neutral-600;
-  cursor: pointer;
-}
+  &:focus {
+    outline: 0;
+  }
 
-.disabled {
-  color: $neutral-400;
-  background-color: inherit;
-  cursor: not-allowed;
+  &.upvoteSelected {
+    color: $primary-500;
+
+    &:hover {
+      color: $primary-500;
+    }
+
+    .Text {
+      color: $primary-500 !important;
+    }
+  }
+
+  &.default {
+    color: $neutral-500;
+  }
+
+  &:hover {
+    background-color: $neutral-100;
+    color: $neutral-600;
+    cursor: pointer;
+
+    .Text {
+      color: $neutral-600;
+    }
+  }
+
+  &.disabled {
+    color: $neutral-400;
+    background-color: inherit;
+    cursor: not-allowed;
+
+    .Text {
+      color: $neutral-400;
+    }
+  }
 }
